feat: update yaml2json workflow
This commit adds a new script, `yaml2json.py`, instead of exec
the code directly as a command-line argument.

The workflow now includes a new filter (`branches`) to ensure it
only triggers on pushes to the main branch when changes are made
to the `QUERIES.yaml` file.

The commit message for push the compiled JSON is changed to:
`build: compile JSON [timestamp] :robot:`, and it includes the
optional prefix `[force]` if the workflow is dispatched manually.

Note: The user email and name are set to ghost for the commit and
push actions, and the target branch for the push is determined by
the `GITHUB_REF_NAME` env variable.